requirements.txt
certifi==2016.9.26
cffi==1.9.1
click==6.6
cryptography==1.6
Flask==0.11.1
future==0.16.0
idna==2.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pyasn1==0.1.9
pycparser==2.17
pyOpenSSL==16.2.0
python-telegram-bot==5.2.0
requests==2.12.3
six==1.10.0
urllib3==1.19.1
Werkzeug==0.11.11
